Split Vulkan raytrace scene renderer
