+++ frame/vulkan/device.cpp
@@ -2647,20 +2647,8 @@ void Device::RecreateSwapchain()
     }
 }
 
-void Device::RecordCommandBuffer(
-    vk::CommandBuffer command_buffer,
-    std::uint32_t image_index)
+SceneState Device::BuildFrameSceneState(vk::Extent2D extent) const
 {
-    vk::CommandBufferBeginInfo begin_info;
-    command_buffer.begin(begin_info);
-
-    const auto extent = swapchain_resources_->GetExtent();
-    const auto& images = swapchain_resources_->GetImages();
-    const auto& render_pass = swapchain_resources_->GetRenderPass();
-    const auto& framebuffers = swapchain_resources_->GetFramebuffers();
-    const auto& gui_render_pass = swapchain_resources_->GetGuiRenderPass();
-    const auto& gui_framebuffers = swapchain_resources_->GetGuiFramebuffers();
-
     const bool has_raytrace_source_meshes =
         level_ && HasRaytracingSourceMeshes(*level_);
     const bool use_shared_transform_hardware_scene =
@@ -2684,505 +2672,19 @@ void Device::RecordCommandBuffer(
                 .GetTemporarySceneRoot();
     }
 
-    const SceneState scene_state =
-        (level_)
-            ? BuildSceneState(
-                  *level_,
-                  frame::Logger::GetInstance(),
-                  {extent.width, extent.height},
-                  elapsed_time_seconds_,
-                  active_program_info_
-                      ? active_program_info_->material_id
-                      : NullId,
-                  !use_compute_raytracing_,
-                  preferred_scene_root,
-                  use_world_space_raytrace_scene)
-            : SceneState{};
-
-    auto update_uniform_buffer = [&](const SceneState& state) {
-        if (!buffer_resources_)
-        {
-            return;
-        }
-        if (!buffer_resources_->GetUniformBuffer())
-        {
-            return;
-        }
-        auto block = MakeUniformBlock(
-            state, elapsed_time_seconds_);
-        buffer_resources_->UpdateUniform(
-            &block, sizeof(UniformBlock));
-    };
-    update_uniform_buffer(scene_state);
-
-    auto transition_output = [&](vk::ImageLayout old_layout,
-                                 vk::ImageLayout new_layout,
-                                 vk::PipelineStageFlags src_stage,
-                                 vk::PipelineStageFlags dst_stage,
-                                 vk::AccessFlags src_access,
-                                 vk::AccessFlags dst_access) {
-        if (!output_image_resources_ ||
-            !output_image_resources_->HasComputeOutputImage())
-        {
-            return;
-        }
-        vk::ImageMemoryBarrier barrier(
-            src_access,
-            dst_access,
-            old_layout,
-            new_layout,
-            VK_QUEUE_FAMILY_IGNORED,
-            VK_QUEUE_FAMILY_IGNORED,
-            output_image_resources_->GetComputeOutputImage(),
-            {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1});
-        command_buffer.pipelineBarrier(
-            src_stage,
-            dst_stage,
-            {},
-            nullptr,
-            nullptr,
-            barrier);
-    };
-
-    if (use_compute_raytracing_ &&
-        descriptor_set_ &&
-        output_image_resources_ &&
-        output_image_resources_->HasComputeOutputImage() &&
-        extent.width > 0 &&
-        extent.height > 0)
+    if (!level_)
     {
-        if (!storage_buffers_ready_ && buffer_resources_)
-        {
-            const auto& storage_buffers =
-                buffer_resources_->GetStorageBuffers();
-            std::vector<vk::BufferMemoryBarrier> buffer_barriers;
-            buffer_barriers.reserve(storage_buffers.size());
-            for (const auto& storage : storage_buffers)
-            {
-                if (!storage.buffer || storage.size == 0)
-                {
-                    continue;
-                }
-                buffer_barriers.emplace_back(
-                    vk::AccessFlagBits::eTransferWrite,
-                    vk::AccessFlagBits::eShaderRead,
-                    VK_QUEUE_FAMILY_IGNORED,
-                    VK_QUEUE_FAMILY_IGNORED,
-                    *storage.buffer,
-                    0,
-                    storage.size);
-            }
-            if (!buffer_barriers.empty())
-            {
-                command_buffer.pipelineBarrier(
-                    vk::PipelineStageFlagBits::eTransfer,
-                    use_raytracing_pipeline_
-                        ? vk::PipelineStageFlagBits::eRayTracingShaderKHR
-                        : vk::PipelineStageFlagBits::eComputeShader,
-                    {},
-                    nullptr,
-                    buffer_barriers,
-                    nullptr);
-            }
-            storage_buffers_ready_ = true;
-        }
-
-        if (output_image_resources_->IsComputeOutputInShaderRead())
-        {
-            transition_output(
-                vk::ImageLayout::eShaderReadOnlyOptimal,
-                vk::ImageLayout::eGeneral,
-                vk::PipelineStageFlagBits::eFragmentShader,
-                use_raytracing_pipeline_
-                    ? vk::PipelineStageFlagBits::eRayTracingShaderKHR
-                    : vk::PipelineStageFlagBits::eComputeShader,
-                vk::AccessFlagBits::eShaderRead,
-                vk::AccessFlagBits::eShaderWrite);
-            output_image_resources_->SetComputeOutputInShaderRead(false);
-        }
-
-        if (use_raytracing_pipeline_ &&
-            pipeline_resources_ &&
-            pipeline_resources_->HasRaytracingPipeline() &&
-            pipeline_resources_->GetRaytracingPipelineLayout())
-        {
-            command_buffer.bindPipeline(
-                vk::PipelineBindPoint::eRayTracingKHR,
-                pipeline_resources_->GetRaytracingPipeline());
-            command_buffer.bindDescriptorSets(
-                vk::PipelineBindPoint::eRayTracingKHR,
-                pipeline_resources_->GetRaytracingPipelineLayout(),
-                0,
-                descriptor_set_,
-                {});
-            command_buffer.traceRaysKHR(
-                pipeline_resources_->GetRaygenSbtRegion(),
-                pipeline_resources_->GetMissSbtRegion(),
-                pipeline_resources_->GetHitSbtRegion(),
-                pipeline_resources_->GetCallableSbtRegion(),
-                extent.width,
-                extent.height,
-                1);
-        }
-        else if (pipeline_resources_ &&
-                 pipeline_resources_->HasComputePipeline() &&
-                 pipeline_resources_->GetComputePipelineLayout())
-        {
-            command_buffer.bindPipeline(
-                vk::PipelineBindPoint::eCompute,
-                pipeline_resources_->GetComputePipeline());
-            command_buffer.bindDescriptorSets(
-                vk::PipelineBindPoint::eCompute,
-                pipeline_resources_->GetComputePipelineLayout(),
-                0,
-                descriptor_set_,
-                {});
-            const std::uint32_t group_x =
-                (extent.width + 7) / 8;
-            const std::uint32_t group_y =
-                (extent.height + 7) / 8;
-            command_buffer.dispatch(group_x, group_y, 1);
-        }
-
-        transition_output(
-            vk::ImageLayout::eGeneral,
-            vk::ImageLayout::eShaderReadOnlyOptimal,
-            use_raytracing_pipeline_
-                ? vk::PipelineStageFlagBits::eRayTracingShaderKHR
-                : vk::PipelineStageFlagBits::eComputeShader,
-            vk::PipelineStageFlagBits::eFragmentShader,
-            vk::AccessFlagBits::eShaderWrite,
-            vk::AccessFlagBits::eShaderRead);
-        output_image_resources_->SetComputeOutputInShaderRead(true);
+        return SceneState{};
     }
-
-    std::array<vk::ClearValue, 1> clear_values{};
-    clear_values[0].color = vk::ClearColorValue(std::array<float, 4>{
-        0.1f,
-        0.1f,
-        0.1f,
-        1.0f});
-
-    auto draw_scene = [&]() -> bool {
-        if (!pipeline_resources_ || !pipeline_resources_->HasGraphicsPipeline())
-        {
-            return false;
-        }
-        command_buffer.bindPipeline(
-            vk::PipelineBindPoint::eGraphics,
-            pipeline_resources_->GetGraphicsPipeline());
-
-        vk::Viewport viewport(
-            0.0f,
-            0.0f,
-            static_cast<float>(extent.width),
-            static_cast<float>(extent.height),
-            0.0f,
-            1.0f);
-        command_buffer.setViewport(0, 1, &viewport);
-
-        vk::Rect2D scissor({0, 0}, extent);
-        command_buffer.setScissor(0, 1, &scissor);
-
-        if (descriptor_set_layout_ && descriptor_set_)
-        {
-            command_buffer.bindDescriptorSets(
-                vk::PipelineBindPoint::eGraphics,
-                pipeline_resources_->GetGraphicsPipelineLayout(),
-                0,
-                descriptor_set_,
-                {});
-        }
-
-        glm::mat4 projection = scene_state.projection;
-        glm::mat4 view = scene_state.view;
-        glm::mat4 model = scene_state.model;
-
-        const bool needs_scene_matrices =
-            pipeline_resources_->GetPushConstantSize() > 0 &&
-            !(pipeline_resources_->UsesProceduralQuadPipeline() &&
-              active_program_info_ &&
-              active_program_info_->uses_time_uniform);
-
-        if (needs_scene_matrices && level_)
-        {
-            try
-            {
-                Camera camera_for_frame(level_->GetDefaultCamera());
-                auto camera_holder_id = level_->GetDefaultCameraId();
-                if (camera_holder_id != NullId)
-                {
-                    auto& node =
-                        level_->GetSceneNodeFromId(camera_holder_id);
-                    auto matrix_node = node.GetLocalModel(
-                        static_cast<double>(elapsed_time_seconds_));
-                    auto inverse_model = glm::inverse(matrix_node);
-                    camera_for_frame.SetFront(
-                        level_->GetDefaultCamera().GetFront() *
-                        glm::mat3(inverse_model));
-                    camera_for_frame.SetPosition(
-                        glm::vec3(
-                            glm::vec4(
-                                level_->GetDefaultCamera().GetPosition(), 1.0f) *
-                            inverse_model));
-                }
-
-                if (extent.height != 0)
-                {
-                    camera_for_frame.SetAspectRatio(
-                        static_cast<float>(extent.width) /
-                        static_cast<float>(extent.height));
-                }
-                projection = camera_for_frame.ComputeProjection();
-                projection[1][1] *= -1.0f;
-                view = camera_for_frame.ComputeView();
-                glm::mat4 rotation = glm::mat4(1.0f);
-                view = rotation * view;
-
-                const auto mesh_pairs =
-                    level_->GetMeshMaterialIds();
-                if (!mesh_pairs.empty())
-                {
-                    auto node_id = mesh_pairs.front().first;
-                    auto& node = level_->GetSceneNodeFromId(node_id);
-                    model = node.GetLocalModel(
-                        static_cast<double>(elapsed_time_seconds_));
-                }
-            }
-            catch (const std::exception& ex)
-            {
-                logger_->warn(
-                    "Failed to compute scene matrices: {}", ex.what());
-            }
-        }
-
-        if (pipeline_resources_->GetPushConstantSize() > 0)
-        {
-            if (pipeline_resources_->UsesProceduralQuadPipeline() &&
-                active_program_info_ &&
-                active_program_info_->uses_time_uniform)
-            {
-                float time = elapsed_time_seconds_;
-                command_buffer.pushConstants(
-                    pipeline_resources_->GetGraphicsPipelineLayout(),
-                    pipeline_resources_->GetPushConstantStages(),
-                    0,
-                    pipeline_resources_->GetPushConstantSize(),
-                    &time);
-            }
-            else
-            {
-                struct alignas(16) PushConstants
-                {
-                    glm::mat4 projection;
-                    glm::mat4 view;
-                    glm::mat4 model;
-                    float time_s;
-                } push_constants{projection, view, model, elapsed_time_seconds_};
-                command_buffer.pushConstants(
-                    pipeline_resources_->GetGraphicsPipelineLayout(),
-                    pipeline_resources_->GetPushConstantStages(),
-                    0,
-                    pipeline_resources_->GetPushConstantSize(),
-                    &push_constants);
-            }
-        }
-
-        if (pipeline_resources_->UsesProceduralQuadPipeline())
-        {
-            command_buffer.draw(6, 1, 0, 0);
-            return true;
-        }
-        else if (mesh_resources_ && !mesh_resources_->Empty())
-        {
-            const auto& mesh = mesh_resources_->GetMeshes().front();
-            const vk::DeviceSize offsets[] = {0};
-            command_buffer.bindVertexBuffers(0, *mesh.vertex_buffer, offsets);
-            if (mesh.index_buffer)
-            {
-                command_buffer.bindIndexBuffer(
-                    *mesh.index_buffer, 0, vk::IndexType::eUint32);
-                command_buffer.drawIndexed(mesh.index_count, 1, 0, 0, 0);
-            }
-            else
-            {
-                command_buffer.draw(mesh.index_count, 1, 0, 0);
-            }
-            return true;
-        }
-        return false;
-    };
-
-    bool scene_pass_executed = false;
-    bool scene_content_rendered = false;
-    if (render_pass && image_index < framebuffers.size())
-    {
-        vk::RenderPassBeginInfo render_pass_info(
-            *render_pass,
-            *framebuffers[image_index],
-            vk::Rect2D({0, 0}, extent),
-            static_cast<std::uint32_t>(clear_values.size()),
-            clear_values.data());
-
-        command_buffer.beginRenderPass(
-            render_pass_info,
-            vk::SubpassContents::eInline);
-        scene_content_rendered = draw_scene();
-        command_buffer.endRenderPass();
-        scene_pass_executed = true;
-    }
-
-    if (!scene_pass_executed && image_index < images.size())
-    {
-        const vk::Image swapchain_image = images[image_index];
-        vk::ImageMemoryBarrier to_transfer_src(
-            {},
-            vk::AccessFlagBits::eTransferRead,
-            vk::ImageLayout::ePresentSrcKHR,
-            vk::ImageLayout::eTransferSrcOptimal,
-            VK_QUEUE_FAMILY_IGNORED,
-            VK_QUEUE_FAMILY_IGNORED,
-            swapchain_image,
-            {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1});
-        command_buffer.pipelineBarrier(
-            vk::PipelineStageFlagBits::eTopOfPipe,
-            vk::PipelineStageFlagBits::eTransfer,
-            {},
-            nullptr,
-            nullptr,
-            to_transfer_src);
-    }
-
-    if (output_image_resources_ &&
-        output_image_resources_->HasSwapchainPreviewImage() &&
-        scene_content_rendered &&
-        image_index < images.size() &&
-        extent.width > 0 && extent.height > 0)
-    {
-        const vk::Image swapchain_image = images[image_index];
-
-        std::array<vk::ImageMemoryBarrier, 2> to_copy_barriers = {
-            vk::ImageMemoryBarrier(
-                vk::AccessFlagBits::eColorAttachmentWrite,
-                vk::AccessFlagBits::eTransferRead,
-                vk::ImageLayout::eTransferSrcOptimal,
-                vk::ImageLayout::eTransferSrcOptimal,
-                VK_QUEUE_FAMILY_IGNORED,
-                VK_QUEUE_FAMILY_IGNORED,
-                swapchain_image,
-                {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1}),
-            vk::ImageMemoryBarrier(
-                output_image_resources_->IsSwapchainPreviewInShaderRead()
-                    ? vk::AccessFlagBits::eShaderRead
-                    : vk::AccessFlags{},
-                vk::AccessFlagBits::eTransferWrite,
-                output_image_resources_->IsSwapchainPreviewInShaderRead()
-                    ? vk::ImageLayout::eShaderReadOnlyOptimal
-                    : vk::ImageLayout::eUndefined,
-                vk::ImageLayout::eTransferDstOptimal,
-                VK_QUEUE_FAMILY_IGNORED,
-                VK_QUEUE_FAMILY_IGNORED,
-                output_image_resources_->GetSwapchainPreviewImage(),
-                {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1})};
-
-        command_buffer.pipelineBarrier(
-            vk::PipelineStageFlagBits::eAllCommands,
-            vk::PipelineStageFlagBits::eTransfer,
-            {},
-            nullptr,
-            nullptr,
-            to_copy_barriers);
-
-        vk::ImageCopy copy_region(
-            {vk::ImageAspectFlagBits::eColor, 0, 0, 1},
-            {0, 0, 0},
-            {vk::ImageAspectFlagBits::eColor, 0, 0, 1},
-            {0, 0, 0},
-            {extent.width, extent.height, 1});
-        command_buffer.copyImage(
-            swapchain_image,
-            vk::ImageLayout::eTransferSrcOptimal,
-            output_image_resources_->GetSwapchainPreviewImage(),
-            vk::ImageLayout::eTransferDstOptimal,
-            copy_region);
-
-        std::array<vk::ImageMemoryBarrier, 2> from_copy_barriers = {
-            vk::ImageMemoryBarrier(
-                vk::AccessFlagBits::eTransferRead,
-                gui_render_callback_
-                    ? vk::AccessFlagBits::eTransferRead
-                    : vk::AccessFlags{},
-                vk::ImageLayout::eTransferSrcOptimal,
-                vk::ImageLayout::eTransferSrcOptimal,
-                VK_QUEUE_FAMILY_IGNORED,
-                VK_QUEUE_FAMILY_IGNORED,
-                swapchain_image,
-                {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1}),
-            vk::ImageMemoryBarrier(
-                vk::AccessFlagBits::eTransferWrite,
-                vk::AccessFlagBits::eShaderRead,
-                vk::ImageLayout::eTransferDstOptimal,
-                vk::ImageLayout::eShaderReadOnlyOptimal,
-                VK_QUEUE_FAMILY_IGNORED,
-                VK_QUEUE_FAMILY_IGNORED,
-                output_image_resources_->GetSwapchainPreviewImage(),
-                {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1})};
-
-        command_buffer.pipelineBarrier(
-            vk::PipelineStageFlagBits::eTransfer,
-            vk::PipelineStageFlagBits::eAllCommands,
-            {},
-            nullptr,
-            nullptr,
-            from_copy_barriers);
-
-        output_image_resources_->SetSwapchainPreviewInShaderRead(true);
-    }
-
-    if (gui_render_callback_ && gui_render_pass &&
-        image_index < gui_framebuffers.size())
-    {
-        vk::RenderPassBeginInfo gui_pass_info(
-            *gui_render_pass,
-            *gui_framebuffers[image_index],
-            vk::Rect2D({0, 0}, extent),
-            0,
-            nullptr);
-        command_buffer.beginRenderPass(
-            gui_pass_info,
-            vk::SubpassContents::eInline);
-        try
-        {
-            gui_render_callback_(command_buffer);
-        }
-        catch (const std::exception& ex)
-        {
-            logger_->error("Failed to render Vulkan GUI: {}", ex.what());
-        }
-        command_buffer.endRenderPass();
-    }
-    else if (image_index < images.size())
-    {
-        const vk::Image swapchain_image = images[image_index];
-        vk::ImageMemoryBarrier to_present(
-            vk::AccessFlagBits::eTransferRead,
-            {},
-            vk::ImageLayout::eTransferSrcOptimal,
-            vk::ImageLayout::ePresentSrcKHR,
-            VK_QUEUE_FAMILY_IGNORED,
-            VK_QUEUE_FAMILY_IGNORED,
-            swapchain_image,
-            {vk::ImageAspectFlagBits::eColor, 0, 1, 0, 1});
-        command_buffer.pipelineBarrier(
-            vk::PipelineStageFlagBits::eTransfer,
-            vk::PipelineStageFlagBits::eBottomOfPipe,
-            {},
-            nullptr,
-            nullptr,
-            to_present);
-    }
-
-    command_buffer.end();
+    return BuildSceneState(
+        *level_,
+        frame::Logger::GetInstance(),
+        {extent.width, extent.height},
+        elapsed_time_seconds_,
+        active_program_info_ ? active_program_info_->material_id : NullId,
+        !use_compute_raytracing_,
+        preferred_scene_root,
+        use_world_space_raytrace_scene);
 }
 
 void Device::CreateGraphicsPipeline()
